mover record_factory and stream_event_data_factory from conftest.py to tests/app/test_stream.py
